# docs/zhenzhengge-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "证证鸽 API",
    "description": "证证鸽后端：案件、证据包、监控任务、通知配置与文书草稿服务。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/cases": {
      "get": {
        "tags": [
          "案件"
        ],
        "summary": "案件列表",
        "operationId": "list_cases_api_v1_cases_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CaseStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "案件状态过滤",
              "title": "Status"
            },
            "description": "案件状态过滤"
          },
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "平台过滤",
              "title": "Platform"
            },
            "description": "平台过滤"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}": {
      "get": {
        "tags": [
          "案件"
        ],
        "summary": "案件详情",
        "operationId": "get_case_api_v1_cases__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/evidence-packs": {
      "get": {
        "tags": [
          "案件"
        ],
        "summary": "案件下证据包列表",
        "operationId": "list_case_evidence_packs_api_v1_cases__case_id__evidence_packs_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaseDetail": {
        "properties": {
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "brand_name": {
            "type": "string",
            "title": "Brand Name"
          },
          "suspect_name": {
            "type": "string",
            "title": "Suspect Name"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "risk_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "evidence_count": {
            "type": "integer",
            "title": "Evidence Count",
            "default": 0
          },
          "template_count": {
            "type": "integer",
            "title": "Template Count",
            "default": 0
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "monitoring_scope": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Monitoring Scope"
          }
        },
        "type": "object",
        "required": [
          "case_id",
          "title",
          "brand_name",
          "suspect_name",
          "platform",
          "risk_score",
          "risk_level",
          "status",
          "updated_at",
          "description"
        ],
        "title": "CaseDetail"
      },
      "CaseListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseSummary"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "CaseListResponse"
      },
      "CaseStatus": {
        "type": "string",
        "enum": [
          "open",
          "monitoring",
          "triaged",
          "drafting",
          "review",
          "archived"
        ],
        "title": "CaseStatus"
      },
      "CaseSummary": {
        "properties": {
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "brand_name": {
            "type": "string",
            "title": "Brand Name"
          },
          "suspect_name": {
            "type": "string",
            "title": "Suspect Name"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "risk_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "case_id",
          "title",
          "brand_name",
          "suspect_name",
          "platform",
          "risk_score",
          "risk_level",
          "status",
          "updated_at"
        ],
        "title": "CaseSummary"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}